+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [25.71.0](https://github.com/bettyblocks/cli/compare/v25.70.0...v25.71.0) (2023-03-14)
+
+
+### Features
+
+* added configs for propname and formatting ([#414](https://github.com/bettyblocks/cli/issues/414)) ([c9b0ef0](https://github.com/bettyblocks/cli/commit/c9b0ef0))
+
+# [25.70.0](https://github.com/bettyblocks/cli/compare/v25.69.0...v25.70.0) (2023-03-14)
+
+
+### Features
+
+* add createNewAction in component options ([#411](https://github.com/bettyblocks/cli/issues/411)) ([8aac99a](https://github.com/bettyblocks/cli/commit/8aac99a))
+
+# [25.69.0](https://github.com/bettyblocks/cli/compare/v25.68.0...v25.69.0) (2023-03-14)
+
+
+### Features
+
+* add optionRef keys ([#412](https://github.com/bettyblocks/cli/issues/412)) ([ef76830](https://github.com/bettyblocks/cli/commit/ef76830))
+
 # [25.68.0](https://github.com/bettyblocks/cli/compare/v25.67.1...v25.68.0) (2023-03-02)